Split out static parts
